Update PlugClean for Gist plugins
In order to install Gist as a plugin, you have to modify its `dir` to be
the `plugin` subdirectory of the normal path of the plugin if it were an
ordinary pathogen-compatible Vim plugin.

    Plug 'https://gist.github.com/952560a43601cd9898f1.git',
        \ { 'dir': g:plug_home.'/xxx/plugin', 'rtp': '..' }

PlugClean command was incompatible with the above trick, so it was
updated not to remove the parent directory of `dir` of each plugin.