git_validate: better error with checkout being ahead/diverged
The `--count` and `--left-right` options seem to be in Git since before
1.7 already.

The error message is improved, and distinguishes between only being
ahead (which is the case if you could push), or being diverged (ahead
and defined).

I think it could be bad to accidentally clean/delete a local checkout
which is ahead, and therefore PlugClean is not mentioned then, which
appears to trigger it to show up in :PlugClean.