Add another install way
Add another install way which vim will download and install the `vim-plug` automatically. Nothing will happen if it was already installed.
The benefit of this approach is that we don't need to run the install command manually. When restoring configurations on a new machine, only thing to do is copy the `vimrc` file or make a symbol link.