Allow pressing 'q' to abort the process; and 'R' to retry
If you press 'q' while `PlugInstall` or `PlugUpdate` is in progress,
the current version of vim-plug immediately closes its window and aborts
all the running processes.

This commit enhances the 'q' map so that it doesn't close the window,
but instead displays aborted tasks so that the user can press 'R' to
retry aborted/skipped tasks.

Fix #1285